requirements.txt
Flask 
flask-login
flask-sqlalchemy
flask-wtf
email_validator
rcon
python-dotenv
pypdf
discord
google-cloud-storage